Move Ralph resilience into recovery
Constraint: Ralph is an interaction recovery subsystem with circuit breakers, outcomes, HITL escalation, and strategy memory.
Rejected: Keep Ralph under src/utils | a domain subsystem inside utilities keeps the architecture misleading.
Confidence: high
Scope-risk: moderate
Directive: Add future Ralph resilience code under src/recovery/ralph and keep tools as callers only.
Tested: npm test -- --runTestsByPath tests/recovery/ralph/circuit-breaker.test.ts tests/recovery/ralph/hitl-escalation.test.ts tests/recovery/ralph/outcome-classifier.test.ts tests/recovery/ralph/ralph-engine.test.ts tests/recovery/ralph/strategy-learner.test.ts tests/recovery/ralph/timeout-budget.test.ts tests/tools/act.test.ts tests/tools/interact.coordinate.test.ts tests/tools/intent.test.ts tests/headed/headed-tool-interop.test.ts tests/vision/vision-find.test.ts tests/e2e/scenarios/domain-memory-persistence.test.ts --runInBand; npm run build; npm run lint; npm run lint:tier; npm run lint:repo-structure; git diff --check
Not-tested: GitHub PR checks not yet run
