+++ src/tools/act.ts
@@ -15,7 +15,7 @@ import { DEFAULT_DOM_SETTLE_DELAY_MS } from '../config/defaults';
 import { normalizeQuery } from '../dom/element-finder';
 import { resolveElementsByAXTree, invalidateAXCache, AXResolvedElement } from '../dom/ax-element-resolver';
 import { getTargetId } from '../cdp/target-id';
-import { classifyOutcome, formatOutcomeLine } from '../utils/ralph/outcome-classifier';
+import { classifyOutcome, formatOutcomeLine } from '../recovery/ralph/outcome-classifier';
 import { humanMouseMove, humanType } from '../stealth/human-behavior';
 import { withTimeout } from '../utils/with-timeout';
 import { cleanupTags, DISCOVERY_TAG } from '../dom/element-discovery';
