+++ src/tools/interact.ts
@@ -17,8 +17,8 @@ import { withTimeout } from '../utils/with-timeout';
 import { makeImageContent } from '../core/screenshot/image-mime';
 import { resolveElementsByAXTree, invalidateAXCache, MATCH_LEVEL_LABELS } from '../dom/ax-element-resolver';
 import { getTargetId } from '../cdp/target-id';
-import { classifyOutcome, formatOutcomeLine } from '../utils/ralph/outcome-classifier';
-import { getCircuitBreaker } from '../utils/ralph/circuit-breaker';
+import { classifyOutcome, formatOutcomeLine } from '../recovery/ralph/outcome-classifier';
+import { getCircuitBreaker } from '../recovery/ralph/circuit-breaker';
 import { humanMouseMove } from '../stealth/human-behavior';
 import { dispatchCoordinateClick } from '../cdp/input';
 import { coerceVerifyMode, runVerify, VERIFY_FIELD_SCHEMA, VerifyReport } from '../core/perception/verify';
